fine

+++ app/src/main/java/uz/texnopos/texnoposedufinance/ui/main/teacher/TeachersInfoFragment.kt
@@ -1,7 +1,9 @@
 package uz.texnopos.texnoposedufinance.ui.main.teacher
 
+import android.app.AlertDialog
 import android.os.Bundle
 import android.view.View
+import android.widget.Toast
 import androidx.lifecycle.Observer
 import androidx.navigation.NavController
 import androidx.navigation.Navigation
@@ -16,7 +18,7 @@ import uz.texnopos.texnoposedufinance.databinding.ActionBarBinding
 import uz.texnopos.texnoposedufinance.databinding.AddActionBarBinding
 import uz.texnopos.texnoposedufinance.databinding.FragmentTeachersInfoBinding
 
-class TeachersInfoFragment: BaseFragment(R.layout.fragment_teachers_info) {
+class TeachersInfoFragment : BaseFragment(R.layout.fragment_teachers_info) {
     lateinit var binding: FragmentTeachersInfoBinding
     lateinit var actBinding: AddActionBarBinding
     private val arg: TeachersInfoFragmentArgs by navArgs()
@@ -32,6 +34,8 @@ class TeachersInfoFragment: BaseFragment(R.layout.fragment_teachers_info) {
 
         val teacherId = arg.teacherId
         setUpObserversDelete()
+        //setUpObservers()
+        //viewModel.getDataCurrentTeacher(teacherId)
 
         actBinding.apply {
             actionBarTitle.text = view.context.getString(R.string.teachers_info)
@@ -41,29 +45,56 @@ class TeachersInfoFragment: BaseFragment(R.layout.fragment_teachers_info) {
             }
             binding.apply {
                 delete.onClick {
-                    viewModel.deleteTeacher(teacherId)
+                    val dialog = AlertDialog.Builder(requireContext())
+                    dialog.setTitle("Удалить")
+                    dialog.setMessage("Вы точно хотите удалить?")
+                    dialog.setIcon(R.drawable.ic_warning)
+                    dialog.setPositiveButton("Да") { _, _ ->
+                        viewModel.deleteTeacher(teacherId)
+                    }
+                    dialog.setNegativeButton("Отмена") { dialog, _ ->
+                        dialog.dismiss()
+                    }
+                    dialog.show()
                 }
+
             }
         }
+    }
 
-    }
-    private fun setUpObserversDelete(){
-        binding.apply {
-            viewModel.deleted.observe(viewLifecycleOwner, Observer {
-                when (it.status) {
-                    ResourceState.LOADING -> loading.visibility(true)
-                    ResourceState.SUCCESS -> {
-                        loading.visibility(false)
-                        toastLN("Успешно удален")
-                    }
-                    ResourceState.ERROR -> {
-                        loading.visibility(false)
-                        toastLN(it.message)
-                    }
-                    else -> {
-                    }
+
+private fun setUpObserversDelete() {
+    binding.apply {
+        viewModel.deleted.observe(viewLifecycleOwner, Observer {
+            when (it.status) {
+                ResourceState.LOADING -> loading.visibility(true)
+                ResourceState.SUCCESS -> {
+                    loading.visibility(false)
+                    toastLN("Успешно удалено")
                 }
-            })
-        }
+                ResourceState.ERROR -> {
+                    loading.visibility(false)
+                    toastLN(it.message)
+                }
+            }
+        })
     }
+}
+
+private fun setUpObservers() {
+    binding.apply {
+        viewModel.current.observe(viewLifecycleOwner, Observer {
+            when (it.status) {
+                ResourceState.LOADING -> loading.visibility(true)
+                ResourceState.SUCCESS -> {
+                    loading.visibility(false)
+                }
+                ResourceState.ERROR -> {
+                    toastLN(it.message)
+                    loading.visibility(false)
+                }
+            }
+        })
+    }
+}
 }
